Port fixes.
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@3196 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/GrammarApplicator.h
@@ -91,8 +91,8 @@ namespace CG3 {
 		int runGrammarOnWindow(Window *window);
 		uint32_t runRulesOnWindow(SingleWindow *current, const int32_t start, const int32_t end);
 
-		Cohort *runSingleTest(SingleWindow *sWindow, uint32_t i, const ContextualTest *test, bool *brk, bool *retval);
-		Cohort *runContextualTest(SingleWindow *sWindow, const uint32_t position, const ContextualTest *test);
+		Cohort *runSingleTest(SingleWindow *sWindow, size_t i, const ContextualTest *test, bool *brk, bool *retval);
+		Cohort *runContextualTest(SingleWindow *sWindow, const size_t position, const ContextualTest *test);
 
 		bool doesTagMatchSet(const uint32_t tag, const uint32_t set);
 		bool doesTagMatchReading(const Reading *reading, const uint32_t tag, bool bypass_index = false);
